Reconstruct runtime module
